// Semester_3_Angular_Personal/Customer/src/app/models/product.model.ts
export class Product{
    productId: number;
    image: string;
    productName: string;
    price: any;
    categoryId: number;
    description: string;
    quantity: number;
    detail: string;
    expireDate: string;
    manufacturerId: number;
    hide: boolean;
    createdAt: string;
    updatedAt: string;
}